pyproject.toml
[tool.poetry]
name = "innisfree"
version = "0.1.0"
description = ""
authors = ["Conor Schaefer <conor.schaefer@gmail.com>"]
include = ["files/*"]

[tool.poetry.scripts]
innisfree = "innisfree.cli:main"

[tool.poetry.dependencies]
python = "^3.7"
pip = "^20.2"
"ruamel.yaml" = "^0.16.10"
requests = "^2.24.0"
configargparse = "^1.2.3"
kopf = "^0.27"
kubernetes = "^11.0.0"
Jinja2 = "^2.11.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
black = "^19.10b0"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
